+++ src/server/devices/baseDeviceInterfaces/iBatteryDevice.ts
@@ -1,0 +1,5 @@
+import { iBaseDevice } from './iBaseDevice';
+
+export interface iBatteryDevice extends iBaseDevice {
+  battery: number;
+}