+++ src/server/devices/hmIPDevices/hmIpDevice.ts
@@ -27,13 +27,10 @@ export class HmIPDevice extends IoBrokerBaseDevice {
     }
 
     switch (idSplit[4]) {
-      case 'OPERATING_VOLTAGE':
-        this.battery = state.val as number;
-        break;
       case 'LOW_BAT':
         const newBatLowVal: boolean = state.val as boolean;
         if (newBatLowVal) {
-          this.log(LogLevel.Alert, `!!BATTERIE FAST LEER!!`);
+          this.log(LogLevel.Warn, `!!BATTERIE FAST LEER!!`);
         }
         break;
     }