+++ src/server/devices/hmIPDevices/hmIpTherm.ts
@@ -2,14 +2,28 @@ import { HmIPDevice } from './hmIpDevice';
 import { DeviceType } from '../deviceType';
 import { LogLevel } from '../../../models';
 import { IoBrokerDeviceInfo } from '../IoBrokerDeviceInfo';
+import { DeviceCapability } from '../DeviceCapability';
+import { iBatteryDevice } from '../baseDeviceInterfaces';
 
-export class HmIpTherm extends HmIPDevice {
+export class HmIpTherm extends HmIPDevice implements iBatteryDevice {
+  public battery: number = -99;
+
   public constructor(pInfo: IoBrokerDeviceInfo) {
     super(pInfo, DeviceType.HmIpTherm);
+    this.deviceCapabilities.push(DeviceCapability.batteryDriven);
   }
 
   public update(idSplit: string[], state: ioBroker.State, initial: boolean = false): void {
     this.log(LogLevel.DeepTrace, `Thermostat Update: ID: ${idSplit.join('.')} JSON: ${JSON.stringify(state)}`);
     super.update(idSplit, state, initial, true);
+    switch (idSplit[3]) {
+      case '0':
+        switch (idSplit[4]) {
+          case 'OPERATING_VOLTAGE':
+            this.battery = ((state.val as number) - 1.8) / 1.2;
+            break;
+        }
+        break;
+    }
   }
 }